+++ NEWS.md
@@ -30,11 +30,13 @@ in the augment method for the chi sq test, .residuals column was renamed to .res
 - Bug fix for tidy.ridgelm returning inconsistent columns (#532)
 - Correct output for  `tidy.mlm(, quick=TRUE)`, add tests (#539 by @MatthieuStigler)
 - `tidy_optim()` provides the standard error if the Hessian is present. (#529 by @billdenney)
-- `tidy.htest()` column names are now run through `make.names()` to ensure syntactic correctness (#549 by @karissawhiting) 
+- Added default arugment of newdata=NULL to augment.prcomp and augment.rqs (#543)
+- Added default arugment of data=NULL to augment.factanal and augment.kmeans (#543)
+- Added default arugments for type.predict and type.residuals to augment.betareg that correspond to default arguments for betareg (#543)
+- Removed coercion warning for augment.factanal (#543)
 - Added method `tidy.lm.beta()` to tidy `lm.beta` class models (#545 by @mattle24)
 - Add feature for glance.biglm to return df.residual
 - Patch bug in glance.lavaan (#577)
-
 
 ## Deprecations